Made state and init ints unsigned short in

+++ src/move.c
@@ -28,8 +28,8 @@ SOFTWARE.
 #define left_def RPI_GPIO_P1_12
 #define right_def RPI_GPIO_P1_13
 
-static int forward_state = 0;
-static int forward_init = 0;
+static unsigned short int forward_state = 0;
+static unsigned short int forward_init = 0;
 
 int forward_set(int arg)
 {
@@ -73,8 +73,8 @@ int forward_status(void)
 	return forward_state;
 }
 
-static int backward_state = 0;
-static int backward_init = 0;
+static unsigned short int backward_state = 0;
+static unsigned short int backward_init = 0;
 
 int backward_set(int arg)
 {
@@ -118,8 +118,8 @@ int backward_status(void)
 }
 
 
-static int left_state = 0;
-static int left_init = 0;
+static unsigned short int left_state = 0;
+static unsigned short int left_init = 0;
 
 int left_set(int arg)
 {
@@ -163,8 +163,8 @@ int left_status(void)
 	return left_state; 
 }
 
-static int right_state = 0;
-static int right_init = 0;
+static unsigned short int right_state = 0;
+static unsigned short int right_init = 0;
 
 int right_set(int arg)
 {
